security_architect_bot/requirements.txt
slack-bolt>=1.16.0
requests>=2.28.0
python-dotenv>=0.19.0
openai>=1.0.0
urllib3>=2.0.0
pytesseract>=0.3.10  # For OCR
Pillow>=10.0.0  # For image processing
google-genai>=1.10.0